+++ crates/matrix-sdk-ui/src/timeline/event_handler.rs
@@ -830,8 +830,12 @@ impl<'a, 'o> TimelineEventHandler<'a, 'o> {
         }
 
         let target = TimelineEventItemId::EventId(redacted.clone());
-        let aggregation =
-            Aggregation::new(self.ctx.flow.timeline_item_id(), AggregationKind::Redaction);
+        let aggregation = Aggregation::new(
+            self.ctx.flow.timeline_item_id(),
+            AggregationKind::Redaction {
+                is_local: false, // We can only get here for remote echoes of redactions.
+            },
+        );
         self.meta.aggregations.add(target.clone(), aggregation.clone());
 
         find_item_and_apply_aggregation(
@@ -951,6 +955,7 @@ impl<'a, 'o> TimelineEventHandler<'a, 'o> {
             forwarder_profile,
             timestamp,
             content,
+            None,
             kind,
             is_room_encrypted,
         );