+++ crates/matrix-sdk-ui/src/timeline/event_item/mod.rs
@@ -78,8 +78,13 @@ pub struct EventTimelineItem {
     pub(super) forwarder_profile: Option<TimelineDetails<Profile>>,
     /// The timestamp of the event.
     pub(super) timestamp: MilliSecondsSinceUnixEpoch,
-    /// The content of the event.
+    /// The content of the event. Might be redacted if a redaction for this
+    /// event is currently being sent or has been received from the server.
     pub(super) content: TimelineItemContent,
+    /// If a redaction for this event is currently being sent but the server
+    /// hasn't yet acknowledged it via its remote echo, the original content
+    /// before redaction. Otherwise, None.
+    pub(super) unredacted_content: Option<TimelineItemContent>,
     /// The kind of event timeline item, local or remote.
     pub(super) kind: EventTimelineItemKind,
     /// Whether or not the event belongs to an encrypted room.
@@ -125,6 +130,7 @@ impl EventTimelineItem {
         forwarder_profile: Option<TimelineDetails<Profile>>,
         timestamp: MilliSecondsSinceUnixEpoch,
         content: TimelineItemContent,
+        unredacted_content: Option<TimelineItemContent>,
         kind: EventTimelineItemKind,
         is_room_encrypted: bool,
     ) -> Self {
@@ -135,6 +141,7 @@ impl EventTimelineItem {
             forwarder_profile,
             timestamp,
             content,
+            unredacted_content,
             kind,
             is_room_encrypted,
         }
@@ -478,7 +485,7 @@ impl EventTimelineItem {
     }
 
     /// Create a clone of the current item, with content that's been redacted.
-    pub(super) fn redact(&self, rules: &RedactionRules) -> Self {
+    pub(super) fn redact(&self, rules: &RedactionRules, is_local: bool) -> Self {
         let content = self.content.redact(rules);
         let kind = match &self.kind {
             EventTimelineItemKind::Local(l) => EventTimelineItemKind::Local(l.clone()),
@@ -491,6 +498,26 @@ impl EventTimelineItem {
             forwarder_profile: self.forwarder_profile.clone(),
             timestamp: self.timestamp,
             content,
+            unredacted_content: is_local.then_some(self.content.clone()),
+            kind,
+            is_room_encrypted: self.is_room_encrypted,
+        }
+    }
+
+    pub(super) fn unredact(&self) -> Self {
+        let Some(content) = &self.unredacted_content else { return self.clone() };
+        let kind = match &self.kind {
+            EventTimelineItemKind::Local(l) => EventTimelineItemKind::Local(l.clone()),
+            EventTimelineItemKind::Remote(r) => EventTimelineItemKind::Remote(r.redact()),
+        };
+        Self {
+            sender: self.sender.clone(),
+            sender_profile: self.sender_profile.clone(),
+            forwarder: self.forwarder.clone(),
+            forwarder_profile: self.forwarder_profile.clone(),
+            timestamp: self.timestamp,
+            content: content.clone(),
+            unredacted_content: None,
             kind,
             is_room_encrypted: self.is_room_encrypted,
         }